+++ tmpl/default.php
@@ -34,9 +34,11 @@ $week = OpeninghoursHelper::getWeek($params);
 		<?php require ModuleHelper::getLayoutPath('mod_openinghours', $params->get('layout', 'default') . '_microdata'); ?>
 	<?php endif; ?>
 
-    <div style="font-size:9px; padding: 5px 10px 5px 0; text-align:center">
-        <a href="https://www.joomill-extensions.com/extensions/opening-hours-days-week-closings" rel="nofollow" target="_blank"><?php echo Text::_('MOD_OPENINGHOURS_FOOTER_CREDIT'); ?></a>
-    </div>
+	<?php if ($params->get('Footer', 1)) : ?>
+        <div class="openinghours-credit" style="font-size:9px; padding: 5px 10px 5px 0; text-align:center">
+            <a href="https://www.joomill-extensions.com/extensions/opening-hours-days-week-closings" rel="nofollow" target="_blank"><?php echo Text::_('MOD_OPENINGHOURS_FOOTER_CREDIT'); ?></a>
+        </div>
+	<?php endif; ?>
 
 	<?php if ($params->get('Debug', 0)) : ?>
         <div class="openinghours-debug" style="font-size:11px; margin-top:10px;">